+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Tests
 - Test `AzureKeyVaultConfig` with `AZURE_KEY_VAULT_FILE` environment variable.
 - Test `NuGetKeyVaultSignTool` in a package.
-- Test `AzureSignTool` in a assembly file.
+- Test `AzureSignTool` in a assembly file, show subject with `GetSignedFileSubject`.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0